+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+- Array coercion: treat JSON `"null"` string as empty array
+- Hash coercion: treat JSON `"null"` string as empty hash
+
 ## [1.16.0] - 2025-02-06
 
 ### Added